refactor: delete cli.rs and put logic in mod.rs
